refac: Split huge test file
