+++ source/databricks/calculation-engine/tests/integration/calculator/calculator_job/configuration.py
@@ -1,0 +1,31 @@
+# Copyright 2020 Energinet DataHub A/S
+#
+# Licensed under the Apache License, Version 2.0 (the "License2");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+executed_batch_id = "0b15a420-9fc8-409a-a169-fbd49479d718"
+energy_supplier_gln_a = "8100000000108"
+energy_supplier_gln_b = "8100000000109"
+balance_responsible_party_gln_a = "1"
+
+
+# Code snippet from https://joelmccune.com/python-dictionary-as-object/
+class DictObj:
+    def __init__(self, in_dict: dict):
+        assert isinstance(in_dict, dict)
+        for key, val in in_dict.items():
+            if isinstance(val, (list, tuple)):
+                setattr(
+                    self, key, [DictObj(x) if isinstance(x, dict) else x for x in val]
+                )
+            else:
+                setattr(self, key, DictObj(val) if isinstance(val, dict) else val)
